requirements.txt
anyio==4.2.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0
Flask==3.0.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==5.1.0
MarkupSafe==2.1.3
orquesta_sdk==2.0.1
pillow==10.2.0
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
PyPDF2==3.0.1
python-docx==1.1.0
python-dotenv==1.0.0
python-pptx==0.6.23
requests==2.31.0
slack-bolt==1.18.1
slack-sdk==3.26.1
sniffio==1.3.0
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
XlsxWriter==3.1.9
